arcticroast/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-allauth==65.12.0
idna==3.10
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
urllib3==2.5.0
